added hacky variable_config
